Seperate repeat code into ReminderGroup
