+++ react-front-end/src/components/Dashboard/Reminders.jsx
@@ -4,6 +4,7 @@ import "semantic-ui-css/semantic.min.css";
 import { Checkbox, Image, Card, Feed } from "semantic-ui-react";
 import wateringcan from "../../assets/wateringcan.png";
 import dayjs from "dayjs";
+import ReminderGroup from "./ReminderGroup";
 const relativeTime = require("dayjs/plugin/relativeTime");
 
 export default function Reminders({ plants, reminders }) {
@@ -18,7 +19,7 @@ export default function Reminders({ plants, reminders }) {
   const remindersWithTime = reminders.map((reminder) => {
     const date1 = dayjs(new Date());
     return  {
-      ...reminder, timeRemaining: reminder.watering_interval - date1.diff(reminder.last_watered, "day")
+      ...reminder, timeRemaining: reminder.watering_interval - date1.diff(reminder.last_watered, "day"), editWatered: () => editWatered(reminder.plant_id)
     }
   })
 
@@ -26,59 +27,7 @@ export default function Reminders({ plants, reminders }) {
   const comingdueReminders = remindersWithTime.filter(element => element.timeRemaining > 0 && element.timeRemaining < 6)
   const notdueReminders = remindersWithTime.filter(element => element.timeRemaining > 6)
 
-  const reminderInstances = reminders.map((reminder) => {
-    console.log("REEEEMINDERS", reminders);
 
-    const date1 = dayjs(new Date());
-    const timeRemaining =
-      reminder.watering_interval - date1.diff(reminder.last_watered, "day");
-    console.log("DATE", date1.diff(reminder.last_watered, "day"));
-    const daysLeft = `${reminder.nickname} in ${timeRemaining} days`;
-
-    if (timeRemaining < 0) {
-      return (
-      <Feed.Event>
-        <Feed.Content>
-          <Feed.Date content="Overdue! Please, please water your baby!" />
-          <Feed.Summary>
-            <Checkbox
-              label={daysLeft}
-              onChange={() => editWatered(reminder.plant_id)}
-            />
-          </Feed.Summary>
-        </Feed.Content>
-      </Feed.Event>
-      )
-    } else if (timeRemaining < 5 && timeRemaining > 0) {
-      return (
-
-              <Feed.Event>
-        <Feed.Content>
-          <Feed.Date content="Coming Soon" />
-          <Feed.Summary>
-            <Checkbox
-              label={daysLeft}
-              onChange={() => editWatered(reminder.plant_id)}
-            />
-          </Feed.Summary>
-        </Feed.Content>
-      </Feed.Event>
-      )
-    } else {
-    return (
-      <Feed.Event>
-        <Feed.Content>
-          {/* <Feed.Date content="Coming Soon" /> */}
-          <Feed.Summary>
-            <Checkbox
-              label={daysLeft}
-              onChange={() => editWatered(reminder.plant_id)}
-            />
-          </Feed.Summary>
-        </Feed.Content>
-      </Feed.Event>
-    );}
-  });
 
   return (
     <Card className="reminders">
@@ -88,7 +37,11 @@ export default function Reminders({ plants, reminders }) {
         </Card.Header>
       </Card.Content>
       <Card.Content>
-        <Feed>{reminderInstances}</Feed>
+        <Feed>
+          <ReminderGroup label={"Overdue! Please water your baby!"} reminders={overdueReminders} />
+          <ReminderGroup label={"Coming due"} reminders={comingdueReminders} />
+          <ReminderGroup label={"Not yet due"} reminders={notdueReminders} />
+        </Feed>
       </Card.Content>
     </Card>
   );
